requirements.txt
numpy==2.2.6
pygame-ce==2.5.6
pymunk==7.2.0
pytest==9.0.2
screeninfo==0.8.1
pygame_gui==0.6.14